Address authorization review round 21 (codex): no token while a challenge URL is unresolved
While a challenge's metadata URL is pending and unresolved the current
authorization server is unknown and no cached token is presented; the
next access retries that URL first (adopt_challenge_metadata /
resolve_pending_challenge) and judges the token by what the document
names.

The challenge-handling helpers (challenge parsing, metadata adoption and
refusal, token retirement on a server change, the peer-advertised URL
checks) move verbatim into OAuthProvider::ChallengeHandling
(lib/mcp_client/auth/oauth_provider/challenge_handling.rb), mixed into
OAuthProvider, to stay within the class length limit; no behaviour
change.
