+++ lib/mcp_client/auth/oauth_provider.rb
@@ -5,6 +5,7 @@ require 'json'
 require 'uri'
 require 'ipaddr'
 require_relative '../auth'
+require_relative 'oauth_provider/challenge_handling'
 
 module MCPClient
   module Auth
@@ -23,6 +24,8 @@ module MCPClient
       # values are consumed by the quoted-string branch, not treated as
       # boundaries. Mirrors HttpTransportBase::AUTH_PARAMS_RUN.
       AUTH_PARAMS_RUN = /\A(?:[\s,]*#{AUTH_PARAM})*/
+
+      include ChallengeHandling
 
       # @!attribute [rw] redirect_uri
       #   @return [String] OAuth redirect URI
@@ -123,6 +126,7 @@ module MCPClient
         # bytes are refused before any binding could attribute them anew.
         return nil if retired_token?(token)
 
+        resolve_pending_challenge
         token = bind_token_issuer(token)
         return nil unless token && token_for_current_issuer?(token)
 
@@ -350,159 +354,6 @@ module MCPClient
 
         logger.debug("OAuth applying authorization header: #{token.to_header[0..20]}...")
         request.headers['Authorization'] = token.to_header
-      end
-
-      # Handle 401 Unauthorized response (for server discovery)
-      # @param response [Faraday::Response] HTTP response
-      # @return [ResourceMetadata, nil] Resource metadata if found
-      def handle_unauthorized_response(response)
-        www_authenticate = response.headers['WWW-Authenticate'] || response.headers['www-authenticate']
-        return nil unless www_authenticate
-
-        # Challenge parameters are read from the Bearer challenge's own
-        # segment only — never from the whole (possibly multi-challenge)
-        # header — so parameters belonging to Basic or another scheme cannot
-        # drive Bearer scope selection or resource metadata discovery. A
-        # header without a Bearer challenge carries no usable Bearer params.
-        bearer_params = bearer_challenge_segment(www_authenticate)
-
-        # MCP 2025-11-25: "Clients MUST treat the scopes provided in the
-        # challenge as authoritative for satisfying the current request" —
-        # including resetting a previously challenged scope when the current
-        # challenge carries none.
-        url = extract_resource_metadata_url(www_authenticate)
-
-        # The challenge header is peer-controlled input: validate the
-        # advertised URL BEFORE storing, fetching, or recording any challenge
-        # state, so a malicious challenge cannot pivot this host into requests
-        # against internal services (SSRF) and cannot leave the provider
-        # holding half of a rejected challenge.
-        validate_peer_advertised_url!(url, 'resource metadata URL (from WWW-Authenticate challenge)') if url
-
-        @challenge_scope = bearer_params && extract_challenge_param(bearer_params, 'scope')
-        return nil unless url
-
-        # Remember the advertised URL even if the fetch below fails, so a
-        # later discovery retries it instead of probing well-known URIs the
-        # challenge already superseded. The current header is the one to
-        # honour: an earlier document, and an earlier refusal, are forgotten
-        # before the fetch so a failed fetch leaves the flow waiting on this
-        # URL rather than completing against stale state.
-        @challenge_metadata_url = url
-        @challenge_resource_metadata = nil
-        @challenge_error = nil
-
-        # This URL was explicitly advertised by the 401 challenge, so a 404 is a
-        # misconfiguration to surface (strict), not a speculative miss to skip.
-        metadata = fetch_resource_metadata(url, strict: true)
-        # Metadata discovery would reject is refused now, whole: it neither
-        # retires the token nor lingers as an authoritative challenge.
-        reject_unacceptable_challenge!(metadata)
-        # A validated challenge supersedes an earlier refused one.
-        @challenge_error = nil
-        # Reuse this challenge-advertised metadata during the subsequent OAuth
-        # flow instead of re-deriving (and possibly missing) the well-known URL.
-        @challenge_resource_metadata = metadata
-        revoke_token_on_authorization_server_change(metadata)
-        metadata
-      end
-
-      # A challenge naming another authorization server than the one the
-      # stored token came from retires that token at once: it is never
-      # presented again, whatever the storage backend can do.
-      # @param resource_metadata [ResourceMetadata]
-      # @return [void]
-      def revoke_token_on_authorization_server_change(resource_metadata)
-        advertised = Array(resource_metadata&.authorization_servers).first
-        known = stored_server_metadata&.issuer
-        return unless advertised && known && advertised != known
-
-        @authorization_server_switched = true
-        # A token another provider sharing the storage already bound to the
-        # advertised server is exactly the token to keep.
-        return if record_bound_to?(stored_token_or_nil, advertised)
-
-        logger.debug('The challenge names another authorization server; retiring the stored token')
-        delete_token(bind_to: Token::RETIRED_ISSUER)
-      end
-
-      # Apply the checks discovery applies to challenge-advertised resource
-      # metadata (resource identity, an acceptable authorization server URL)
-      # before anything acts on it; a failing document refuses the whole
-      # challenge (see {#reject_challenge!}).
-      # @param resource_metadata [ResourceMetadata]
-      # @return [void]
-      # @raise [MCPClient::Errors::ConnectionError] when the challenge is refused
-      def reject_unacceptable_challenge!(resource_metadata)
-        begin
-          validate_resource_matches!(resource_metadata)
-        rescue MCPClient::Errors::ConnectionError => e
-          reject_challenge!(e.message)
-        end
-        advertised = Array(resource_metadata.authorization_servers).first
-        reject_challenge!('Protected resource metadata does not advertise any authorization_servers') unless advertised
-
-        validate_peer_advertised_url!(advertised, 'authorization server URL (from resource metadata)')
-      end
-
-      # Extract the protected-resource-metadata URL from a WWW-Authenticate header.
-      # Per RFC 9728 the parameter is `resource_metadata`; a legacy `resource`
-      # parameter is accepted as a fallback for older servers. Only the Bearer
-      # challenge's own segment is consulted, so a parameter belonging to
-      # another scheme's challenge can never drive discovery.
-      # @param header [String] the WWW-Authenticate header value
-      # @return [String, nil] the metadata URL if present
-      def extract_resource_metadata_url(header)
-        params = bearer_challenge_segment(header)
-        return nil unless params
-
-        # Auth-params may include optional whitespace around '=' (RFC 7235).
-        # Quoted form: resource_metadata = "https://..."
-        if (m = params.match(/resource_metadata\s*=\s*"([^"]+)"/))
-          return m[1]
-        end
-
-        # Unquoted token form: resource_metadata = https://...
-        if (m = params.match(/resource_metadata\s*=\s*([^,\s]+)/))
-          return m[1]
-        end
-
-        # Legacy fallback: resource="https://..."
-        params.match(/resource\s*=\s*"([^"]+)"/)&.captures&.first
-      end
-
-      # Extract the Bearer challenge's own parameter segment from a (possibly
-      # multi-challenge) WWW-Authenticate header, so params belonging to other
-      # schemes (e.g. `Basic resource_metadata="...", Bearer realm="x"`) are
-      # never attributed to the Bearer challenge. Mirrors
-      # HttpTransportBase#bearer_challenge_segment.
-      # @param header [String, nil] the WWW-Authenticate header value
-      # @return [String, nil] the Bearer challenge's parameters (possibly
-      #   empty), or nil when the header has no Bearer challenge
-      def bearer_challenge_segment(header)
-        return nil unless header
-
-        # Locate the Bearer scheme token only OUTSIDE quoted strings: a
-        # quoted value such as realm="prefix Bearer x" must not anchor the
-        # segment.
-        masked = header.gsub(/"(?:\\.|[^"\\])*"/) { |q| "\"#{' ' * (q.length - 2)}\"" }
-        match = masked.match(/(?:\A|[\s,])Bearer(?=[\s,]|\z)/i)
-        return nil unless match
-
-        header[match.end(0)..][AUTH_PARAMS_RUN]
-      end
-
-      # Extract an auth-param value from a WWW-Authenticate header
-      # (quoted or unquoted form, optional whitespace around '=').
-      # @param header [String] the WWW-Authenticate header value
-      # @param name [String] the auth-param name
-      # @return [String, nil] the parameter value if present
-      def extract_challenge_param(header, name)
-        if (m = header.match(/(?:^|[\s,])#{Regexp.escape(name)}\s*=\s*"([^"]*)"/i))
-          return m[1]
-        end
-
-        header.match(/(?:^|[\s,])#{Regexp.escape(name)}\s*=\s*([^,\s]+)/i)&.captures&.first
       end
 
       # Scope requested by the most recent WWW-Authenticate challenge.
@@ -1007,72 +858,6 @@ module MCPClient
         enforce_https!(server_metadata.registration_endpoint, 'registration endpoint')
       end
 
-      # Validate a URL that a peer advertised to us (a 401 challenge's
-      # resource_metadata, or PRM authorization_servers).
-      #
-      # Stricter than enforce_https!, which exists for URLs the OPERATOR
-      # configured and therefore tolerates plain-HTTP loopback for local
-      # development. Applying that exception to peer-supplied input would
-      # leave the reported SSRF intact against the most sensitive targets of
-      # all — services listening only on localhost. The loopback exception is
-      # honored here only when the configured MCP server is itself loopback,
-      # i.e. the developer is already pointed at a local stack.
-      #
-      # The rejection is recorded so a later discovery fails closed instead of
-      # silently reusing cached authorization-server metadata.
-      #
-      # NOTE: hostnames are checked literally. This does not resolve DNS, so a
-      # public name that resolves to a private address is not caught here;
-      # that needs resolution-time checking in the HTTP layer.
-      # @param url [String] the peer-advertised URL
-      # @param label [String] human-readable name for errors
-      # @raise [MCPClient::Errors::ConnectionError] if the URL is not acceptable
-      def validate_peer_advertised_url!(url, label)
-        uri = URI.parse(url)
-        host = uri.hostname.to_s.downcase
-
-        if uri.scheme != 'https' && !(uri.scheme == 'http' && local_development?)
-          reject_challenge!("OAuth #{label} must use HTTPS: #{safe_error_text(url)}")
-        end
-        reject_challenge!("OAuth #{label} must not target a loopback or private address: #{safe_error_text(url)}") if
-          local_address?(host) && !local_development?
-      rescue URI::InvalidURIError
-        reject_challenge!("OAuth #{label} is not a valid URL: #{safe_error_text(url)}")
-      end
-
-      # @param message [String] why the challenge was refused
-      # @raise [MCPClient::Errors::ConnectionError] always
-      def reject_challenge!(message)
-        # Drop every scrap of the refused challenge so nothing half-applied
-        # survives, and remember why for the next discovery attempt.
-        @challenge_scope = nil
-        @challenge_metadata_url = nil
-        @challenge_resource_metadata = nil
-        @resource_metadata = nil
-        @challenge_error = message
-        raise MCPClient::Errors::ConnectionError, message
-      end
-
-      # @return [Boolean] whether the configured MCP server is itself local,
-      #   in which case local discovery targets are expected
-      def local_development?
-        local_address?(URI.parse(server_url).hostname.to_s.downcase)
-      rescue URI::InvalidURIError
-        false
-      end
-
-      # @param host [String] a downcased hostname
-      # @return [Boolean] whether it names a loopback, private or link-local address
-      def local_address?(host)
-        return true if %w[localhost 127.0.0.1 ::1 0.0.0.0].include?(host)
-        return true if host.end_with?('.localhost', '.local', '.internal')
-        return true if host.start_with?('127.', '10.', '192.168.', '169.254.')
-        return true if host.match?(/\A172\.(1[6-9]|2\d|3[01])\./)
-        return true if host.match?(/\A\[?(fc|fd|fe80)/)
-
-        false
-      end
-
       # @param url [String, nil] endpoint URL
       # @param label [String] human-readable endpoint name for errors
       # @raise [MCPClient::Errors::ConnectionError] if the URL is not HTTPS (non-localhost)
@@ -1416,6 +1201,8 @@ module MCPClient
       def current_issuer_for_tokens
         advertised = Array(@challenge_resource_metadata&.authorization_servers).first
         return advertised if advertised.is_a?(String)
+        # An unresolved challenge URL means the current server is unknown.
+        return nil if @challenge_metadata_url
 
         stored_server_metadata&.issuer
       end
